requirements.txt
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
Flask==3.0.2
Flask-Session==0.6.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
mypy-extensions==1.0.0
nylas==6.1.0
packaging==23.2
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1